requirements.txt
python-telegram-bot==20.7
python-telegram-bot[webhooks]==20.7
py-clob-client==0.25.0
py-builder-signing-sdk>=0.0.2
py-builder-relayer-client>=0.0.1
python-dotenv==1.0.0
requests==2.31.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1
redis==5.0.1
web3>=7.0.0
aiohttp==3.9.1
APScheduler==3.10.4
cryptography>=41.0.0
eth-account>=0.10.0
asyncpg==0.29.0
greenlet==3.3.1
prometheus_client>=0.20.0
aiosqlite
privy-client

